Don't allow removal of badge scanners that have already scanned badges

+++ postgresqleu/confsponsor/scanning.py
@@ -66,8 +66,11 @@ def sponsor_scanning(request, sponsorid):
                     try:
                         scanner = SponsorScanner.objects.get(sponsor=sponsor, pk=rid)
                         n = scanner.scanner.fullname
-                        scanner.delete()
-                        messages.info(request, "Attendee {0} removed from scanning".format(n))
+                        if scanner.scanner.scanned_attendees.exists():
+                            messages.warning(request, "Attende {0} has scanned badges already, cannot be removed".format(n))
+                        else:
+                            scanner.delete()
+                            messages.info(request, "Attendee {0} removed from scanning".format(n))
                     except SponsorScanner.DoesNotExist:
                         messges.error(request, "Attendee not found")
                     return HttpResponseRedirect(".")
